+++ game/var_global.h
@@ -1,16 +1,17 @@
 #ifndef GLOBAL_H
 #define GLOBAL_H
 
-/* 全局筹码 */
-extern int pot;
-extern int chip;
-extern int chip1;
+/* 全局筹码，使用 LONG LONG 防止溢出 */
+#include <stdint.h>
+extern long long pot;
+extern long long chip;
+extern long long chip1;
 
-void set_chip(int value);
-int get_chip(void);
-void set_chip1(int value);
-int get_chip1(void);
-void set_pot(int value);
-int get_pot(void);
+void set_chip(long long value);
+long long get_chip(void);
+void set_chip1(long long value);
+long long get_chip1(void);
+void set_pot(long long value);
+long long get_pot(void);
 
 #endif /* GLOBAL_H */
